[test][int][ctrl][jobs][deployment] Move webhook tests.
